+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: 'ConfigSchema' type init
 - feat: `od.Highlights.*` type added
 - refactor: `config` module
+- refactor: `config.config` table renamed to `config.schema`
 
 ### Changed